supplier edit delete

+++ src/app/customerinvoiceintermediatedisplay/customerinvoiceintermediatedisplay.component.ts
@@ -39,6 +39,7 @@ export class CustomerinvoiceintermediatedisplayComponent implements OnInit {
   name="";
   i=0;
   totalmamount=0;
+  displaycustomerorsupplier="NONE";
   generatePDF(action = 'open') {
     let docDefinition = {
       content: [
@@ -152,40 +153,85 @@ export class CustomerinvoiceintermediatedisplayComponent implements OnInit {
     var id=this.sharedservice.getidforcustomeredit();
     var status=this.sharedservice.getcustomerinvoicestatus();
     this.status=status;
-    if(status=="approved"){
-      this.api.getCustomerInvoioceFromId(id).subscribe((data)=>{     
-        this.api.getCustomerNameFromId(data[0].customerid).subscribe((customername:any)=>{        
-         this.name=customername.name;
-         this.invoice.date=data[0].date;
-         this.invoice.duedate=data[0].duedate;
-         this.invoice.invoiceno=data[0].invoiceid;
-         this.invoice.referenceno=data[0].reference;
-         this.invoice.additionalDetails=data[0].additionaldetails;  
-         this.totalmamount=data[0].totalamount;
-         for(var i=0;i<data[0].products.length;i++)  {
-          this.invoice.products.push(new Product());
-           this.invoice.products[i]=data[0].products[i];
-         }  
-        })
-      });
+    if(this.sharedservice.getCustomerOrSupplier()=="Customer"){
+      this.displaycustomerorsupplier="Customer"; 
+      if(status=="approved"){
+        this.api.getCustomerInvoioceFromId(id).subscribe((data)=>{     
+          this.api.getCustomerNameFromId(data[0].customerid).subscribe((customername:any)=>{        
+           this.name=customername.name;
+           this.invoice.date=data[0].date;
+           this.invoice.duedate=data[0].duedate;
+           this.invoice.invoiceno=data[0].invoiceid;
+           this.invoice.referenceno=data[0].reference;
+           this.invoice.additionalDetails=data[0].additionaldetails;  
+           this.totalmamount=data[0].totalamount;
+           for(var i=0;i<data[0].products.length;i++)  {
+            this.invoice.products.push(new Product());
+             this.invoice.products[i]=data[0].products[i];
+           }  
+          })
+        });
+      }
+      else if(status=="draft") {
+        this.api.getDraftCustomerInvoioceFromId(id).subscribe((data)=>{
+          this.api.getCustomerNameFromId(data[0].customerid).subscribe((customername:any)=>{        
+            this.name=customername.name;
+            this.invoice.date=data[0].date;
+            this.invoice.duedate=data[0].duedate;
+            this.invoice.invoiceno=data[0].invoiceid;
+            this.invoice.referenceno=data[0].reference;
+            this.invoice.additionalDetails=data[0].additionaldetails;  
+            this.totalmamount=data[0].totalamount;  
+            for(var i=0;i<data[0].products.length;i++)  {
+             this.invoice.products.push(new Product());
+              this.invoice.products[i]=data[0].products[i];
+            }  
+           })
+        });
+      }
     }
-    else if(status=="draft") {
-      this.api.getDraftCustomerInvoioceFromId(id).subscribe((data)=>{
-        this.api.getCustomerNameFromId(data[0].customerid).subscribe((customername:any)=>{        
-          this.name=customername.name;
-          this.invoice.date=data[0].date;
-          this.invoice.duedate=data[0].duedate;
-          this.invoice.invoiceno=data[0].invoiceid;
-          this.invoice.referenceno=data[0].reference;
-          this.invoice.additionalDetails=data[0].additionaldetails;  
-          this.totalmamount=data[0].totalamount;  
-          for(var i=0;i<data[0].products.length;i++)  {
-           this.invoice.products.push(new Product());
-            this.invoice.products[i]=data[0].products[i];
-          }  
-         })
-      });
+     else if(this.sharedservice.getCustomerOrSupplier()=="Supplier"){
+      this.displaycustomerorsupplier="Supplier"; 
+      if(status=="approved"){
+        this.api.getSupplierInvoioceFromId(id).subscribe((data)=>{     
+          this.api.getSupplierNameFromId(data[0].customerid).subscribe((customername:any)=>{        
+           this.name=customername.name;
+           this.invoice.date=data[0].date;
+           this.invoice.duedate=data[0].duedate;
+           this.invoice.invoiceno=data[0].invoiceid;
+           this.invoice.referenceno=data[0].reference;
+           this.invoice.additionalDetails=data[0].additionaldetails;  
+           this.totalmamount=data[0].totalamount;
+           for(var i=0;i<data[0].products.length;i++)  {
+            this.invoice.products.push(new Product());
+             this.invoice.products[i]=data[0].products[i];
+           }  
+          })
+        });
+      }
+      else if(status=="draft") {
+        this.api.getDraftSupplierInvoioceFromId(id).subscribe((data)=>{
+          this.api.getSupplierNameFromId(data[0].customerid).subscribe((customername:any)=>{        
+            this.name=customername.name;
+            this.invoice.date=data[0].date;
+            this.invoice.duedate=data[0].duedate;
+            this.invoice.invoiceno=data[0].invoiceid;
+            this.invoice.referenceno=data[0].reference;
+            this.invoice.additionalDetails=data[0].additionaldetails;  
+            this.totalmamount=data[0].totalamount;  
+            for(var i=0;i<data[0].products.length;i++)  {
+             this.invoice.products.push(new Product());
+              this.invoice.products[i]=data[0].products[i];
+            }  
+           })
+        });
+      }
     }
+    else{    
+          this.displaycustomerorsupplier="NONE"; 
+          this.router.navigate(['/report']);     
+    }   
+
    }
 
   ngOnInit(): void {
@@ -210,19 +256,40 @@ export class CustomerinvoiceintermediatedisplayComponent implements OnInit {
   }
   deleteInvoice(){  
     var id=this.sharedservice.getidforcustomeredit();
-    if(this.status=="approved"){
-      this.api.deleteCustomerInvoice(id).subscribe((data:any)=>{
-        window.alert(data.msg);
-        this.router.navigate(['/displaycustomerinvoices']);          
-      });
-      
+    if(this.sharedservice.getCustomerOrSupplier()=="Customer"){
+      if(this.status=="approved"){
+        this.api.deleteCustomerInvoice(id).subscribe((data:any)=>{
+          window.alert(data.msg);
+          this.router.navigate(['/displaycustomerinvoices']);          
+        });        
+      }
+      else if(this.status=="draft") {
+        this.api.deleteCustomerInvoiceFromDraft(id).subscribe((data:any)=>{
+          window.alert(data.msg);
+          this.router.navigate(['/displaycustomerinvoices']);          
+        });
+      }
     }
-    else if(this.status=="draft") {
-      this.api.deleteCustomerInvoiceFromDraft(id).subscribe((data:any)=>{
-        window.alert(data.msg);
-        this.router.navigate(['/displaycustomerinvoices']);          
-      });
+     else if(this.sharedservice.getCustomerOrSupplier()=="Supplier"){
+      if(this.status=="approved"){
+        this.api.deleteSupplierInvoice(id).subscribe((data:any)=>{
+          window.alert(data.msg);
+          this.router.navigate(['/displaycustomerinvoices']);          
+        });
+        
+      }
+      else if(this.status=="draft") {
+        this.api.deleteSupplierInvoiceFromDraft(id).subscribe((data:any)=>{
+          window.alert(data.msg);
+          this.router.navigate(['/displaycustomerinvoices']);          
+        });
+      }
     }
+    else{    
+          this.displaycustomerorsupplier="NONE"; 
+          this.router.navigate(['/report']);     
+    }
+   
    
   }
   setasCustomer(){
